Modularize Form model and its related test cases
