+++ app/models/organization.rb
@@ -20,7 +20,7 @@ class Organization < ApplicationRecord
   has_many :users, through: :staff_accounts
   has_many :pets
   has_many :default_pet_tasks
-  has_many :forms, dependent: :destroy
+  has_many :forms, class_name: "CustomForm::Form", dependent: :destroy
   has_many :faqs
 
   has_one :profile, dependent: :destroy, class_name: "OrganizationProfile", required: true
