+++ app/models/question.rb
@@ -27,7 +27,7 @@
 #  fk_rails_...  (form_id => forms.id)
 #
 class Question < ApplicationRecord
-  belongs_to :form
+  belongs_to :form, class_name: "CustomForm::Form"
 
   has_one :organization, through: :form
