+++ app/views/organizations/staff/custom_form/questions/_form.html.erb
@@ -1,6 +1,6 @@
 <div class="row">
   <div class="col-md-8 mx-auto card p-5 bigger">
-    <%= bootstrap_form_with model: [:staff, :custom_form, form, question] do |f| %>
+    <%= bootstrap_form_with model: [:staff, form, question] do |f| %>
       <div class="form-group">
         <%= f.text_field :name, class: 'form-control' %>
       </div>
