+++ test/factories/custom_form/forms.rb
@@ -1,5 +1,5 @@
 FactoryBot.define do
-  factory :form do
+  factory :form, class: "custom_form/form" do
     name { Faker::Lorem.sentence }
     description { Faker::Lorem.sentence }
     title { name }
